modify 홈화면 지도 리렌더링방식 변경, 집방형식과 같게함
