+++ client/src/components/DefaultMap/style.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+export const Styled = {
+  Map: styled.div`
+    margin: 5px;
+
+    width: 100%;
+    height: 99%;
+    border-radius: 10px;
+  `,
+  MapBox: styled.div`
+    width: auto;
+  `,
+};
