+++ client/src/components/PlaceList/PlaceList.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, useCallback } from "react";
-import { useSetRecoilState, useRecoilValue } from "recoil";
+import { useSetRecoilState, useRecoilValue, useRecoilValueLoadable } from "recoil";
 import { Styled } from "./style";
 import { MemoCards } from "../PlaceCard/PlaceCards";
 
@@ -23,12 +23,14 @@ function PlaceList() {
   const accessToken = useRecoilValue(token);
   const kakaoToken = useRecoilValue(kToken);
   const placeList = useRecoilValue(placelist);
+  // const visitedList = useRecoilValueLoadable(placelist);
   const setPlaceLocation = useSetRecoilState(placelocation);
   const setPlaceAddress = useSetRecoilState(placeaddress);
   const setImgURL = useSetRecoilState(placeimg);
   const setTitle = useSetRecoilState(placetitle);
   const [ScrollY, setScrollY] = useState(0);
   const [btnStatus, setBtnStatus] = useState(false);
+  // const placeList = visitedList.contents;
 
   //! Top 버튼에 필요한 주석
   useEffect(() => {
@@ -71,7 +73,7 @@ function PlaceList() {
     setTitle(title);
     setPlaceAddress(address);
   }
-  console.log(ableToSearchPlace);
+
   return (
     <>
       <Styled.PlaceLists>
