+++ client/src/index.js
@@ -13,6 +13,7 @@ ReactDOM.render(
     <RecoilRoot>
       <BrowserRouter>
         <App />
+
         <ToastContainer />
       </BrowserRouter>
     </RecoilRoot>
