+++ client/src/recoil/recoil.js
@@ -175,7 +175,7 @@ export const isClickedNowLocation = atom({
 
 // ! 위치기반 API - Homemap.js랑 연결
 export const getWTM = selector({
-  key: "getWTN",
+  key: "getWTM",
   get: async ({ get }) => {
     const result = await axios.get(
       // `https://dapi.kakao.com/v2/local/geo/transcoord.json?x=${get(pickpoint)[1]}&y=${
@@ -197,6 +197,46 @@ export const getWTM = selector({
     return data;
   },
 });
+
+// export const placelist = selector({
+//   key: "placelist",
+//   get: async ({ get }) => {
+//     const wtm = get(getWTM);
+//     const result = await axios
+//       .get(`${process.env.REACT_APP_API_URL}/home`, {
+//         headers: {
+//           // Authorization: `Bearer ${accessToken || kakaoToken}`,
+//           "Content-Type": "application/json",
+//         },
+//         params: {
+//           areacode: "null",
+//           sigungucode: "null",
+//           radius: 10000,
+//           clientwtmx: wtm.x,
+//           clientwtmy: wtm.y,
+//           tag: "null", //null로 넣으면 undefined되어서, 문자열로 넣겠음
+//           searchWord: "null",
+//         },
+//         withCredentials: true,
+//       })
+//       .then((res) => {
+//         const list = res.data.data.map((el) => {
+//           return [
+//             Number(el.post_mapy),
+//             Number(el.post_mapx),
+//             el.post_title,
+//             el.post_firstimage,
+//             el.post_addr1,
+//             el.post_contentid,
+//           ];
+//         });
+
+//         return list;
+//       });
+
+//     return result;
+//   },
+// });
 
 //! myVisited 모달
 export const visitedModal = atom({
